reorganizaçao do spa
