+++ src/index.html
@@ -11,12 +11,6 @@
 
 <body>
     <main class="formRegister">
-        <nav>
-            <ul>
-                <li id='login'>Login</li>
-                <li id='register'>Cadastre-se</li>
-            </ul>
-        </nav>
         <div id="root"></div>
     </main>
